complete ingestion extraction and shared utility consolidation
