+++ proj/260210-stack-integrated/review/12-module-refactoring.md
@@ -78,10 +78,10 @@ This plan ensures complete resolution of all 5 critical issues:
 
 **Fix:**
 
-- [ ] Create `src/quilt_mcp/types/common.py` with shared type definitions
-- [ ] Create `src/quilt_mcp/context/utils.py` for context-specific utilities
-- [ ] Move non-context utilities to `src/quilt_mcp/utils/helpers.py`
-- [ ] Update imports:
+- [x] Create `src/quilt_mcp/types/common.py` with shared type definitions
+- [x] Create `src/quilt_mcp/context/utils.py` for context-specific utilities
+- [x] Move non-context utilities to `src/quilt_mcp/utils/helpers.py`
+- [x] Update imports:
   - `context/handler.py`: import from `types.common`
   - `context/factory.py`: import from `types.common`
   - `services/workflow_service.py`: import from `context.utils`
@@ -587,7 +587,7 @@ platform_backend.Platform_Backend
       # Future enhancement
   ```
 
-- [ ] Move `package_create_from_s3()` (after Phase 3 refactoring)
+- [x] Move `package_create_from_s3()` (after Phase 3 refactoring)
 - [x] Update MCP tool registration to import from new module
 
 **Expected Impact:** ~200 lines moved from packages.py
@@ -826,7 +826,7 @@ def some_operation(self, ...):
 
 **Problem:** Same utility functions repeated across multiple backends.
 
-- [ ] Create `src/quilt_mcp/backends/utils.py`:
+- [x] Create `src/quilt_mcp/backends/utils.py`:
 
   ```python
   """Shared utilities for backend implementations."""
@@ -847,7 +847,7 @@ def some_operation(self, ...):
       # Currently duplicated in backends
   ```
 
-- [ ] Find all duplicated utility functions:
+- [x] Find all duplicated utility functions:
 
   ```bash
   # Search for common patterns
@@ -856,8 +856,8 @@ def some_operation(self, ...):
   grep -rn "def.*build.*s3" src/quilt_mcp/backends/
   ```
 
-- [ ] Extract to `backends/utils.py`
-- [ ] Update all backends to import from utils
+- [x] Extract to `backends/utils.py`
+- [x] Update all backends to import from utils
 
 **Expected Impact:** -50 to -100 lines total across backends
 
@@ -1211,18 +1211,18 @@ def update_package_revision(
 
 ### 8.4 Update Documentation
 
-- [ ] Update `ARCHITECTURE.md`:
+- [x] Update `ARCHITECTURE.md`:
   - [ ] Document new module structure
   - [ ] Document Template Method pattern usage
   - [ ] Document GraphQL client abstraction
   - [ ] Add diagrams for complex workflows
 
-- [ ] Update `CONTRIBUTING.md`:
+- [x] Update `CONTRIBUTING.md`:
   - [ ] Add guidelines on module size
   - [ ] Add guidelines on avoiding circular imports
   - [ ] Add examples of good code organization
 
-- [ ] Update inline documentation:
+- [x] Update inline documentation:
   - [ ] Ensure new modules have clear docstrings
   - [ ] Ensure refactored functions have updated docs
   - [ ] Remove outdated comments
@@ -1231,7 +1231,7 @@ def update_package_revision(
 
 ### 8.5 Create Migration Guide
 
-- [ ] Create `docs/refactoring/MIGRATION_GUIDE.md`:
+- [x] Create `docs/refactoring/MIGRATION_GUIDE.md`:
 
   ```markdown
   # Module Refactoring Migration Guide
@@ -1283,7 +1283,7 @@ def update_package_revision(
 - [ ] **448-line function** → Refactored to ~100 lines with extracted modules
 - [x] **Platform backend orchestration duplication** → Removed, uses base class
 - [x] **15 circular import cycles** → Eliminated, 0 cycles remain
-- [ ] **Mixed concerns in tools/packages.py** → Separated by responsibility
+- [x] **Mixed concerns in tools/packages.py** → Separated by responsibility
 - [x] **Scattered GraphQL operations** → Abstracted into client
 
 ### Code Quality Metrics
