+++ src/quilt_mcp/backends/utils.py
@@ -1,0 +1,11 @@
+"""Shared utilities for backend implementations."""
+
+from __future__ import annotations
+
+from quilt_mcp.utils.helpers import build_s3_key, extract_bucket_from_registry, normalize_s3_registry
+
+__all__ = [
+    "extract_bucket_from_registry",
+    "normalize_s3_registry",
+    "build_s3_key",
+]
