+++ src/quilt_mcp/types/__init__.py
@@ -1,0 +1,5 @@
+"""Shared type aliases for cross-module use."""
+
+from .common import JsonDict, JsonMap, MutableJsonMap, StringMap
+
+__all__ = ["JsonDict", "JsonMap", "MutableJsonMap", "StringMap"]
